Fix linter erors
