--- .github/workflows/app/student.rb
@@ -1,16 +1,0 @@
-require "./person.rb"
-
-class Student < Person
-
-  attr_reader :classroom
-  attr_writer :classroom
-
-  def initialize(classroom,id, age, name, parent_permission)
-    super(id, age, name, parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    return "¯\(ツ)/¯"
-  end
-end
